\name{fev}
\alias{fev}
\docType{data}
\title{
Modified Rosner's FEV data
}
\description{
Modified children's FEV data. FEV was replaced by log(FEV). Age 3 was grouped to age 4 and Age 19+ was grouped to age 18.   
}
\usage{data(fev)}
\format{
  A data frame with 654 observations on the following 5 variables.
  \describe{
    \item{\code{age}}{a numeric vector}
    \item{\code{fev}}{a numeric vector}
    \item{\code{height}}{a numeric vector}
    \item{\code{male}}{a numeric vector}
    \item{\code{smoke}}{a numeric vector}
  }
}

\references{
Rosner, B. (1999), Fundamentals of Biostatistics, 5th ed., Pacific Grove, CA: Duxbury. 
}

\keyword{datasets}
